+++ CHANGELOG.md
@@ -22,6 +22,22 @@ All notable changes to this project will be documented here. For more details, v
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+
+### Added
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+- Some false positive DDoS-Guard detections
+
+### Security
+
 ## [9.13.0] - 2026-05-29
 
 ### Changed